Refactor: Extract new mod `binstall/install.rs`
